DBZ-8986 Additional refactors
